+++ src/Model/Watch/EventTypes.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace TK\GitHubWebhook\Model\Watch;
+
+enum EventTypes: string
+{
+    case STARTED = "started";
+}